+++ eltakobus/device.py
@@ -982,7 +982,34 @@ class F4HK14(FHK14, HasProgrammableRPS):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-known_objects = [FAM14, FUD14, FUD14_800W, FSB14, FSR14_1x, FSR14_2x, FSR14_4x, F4SR14_LED, F3Z14D, FMZ14, FWG14MS, FSU14, FMSR14, FWZ14_65A, FSG14_1_10V, FGW14_USB, FDG14, FHK14, F4HK14, FAE14SSR]
+
+class FTD14(BusObject):
+    size=1
+    discovery_names = [ bytes((0x04, 0xa0)) ]
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    async def get_base_id(self) -> str:
+        """Gets base id from memory."""
+        mem_line = await self.read_mem_line(1)
+        return b2s(mem_line[0:4])
+    
+    async def get_base_id_in_bytes(self) -> bytes:
+        """Gets base id from memory."""
+        mem_line = await self.read_mem_line(1)
+        return mem_line[0:4]
+
+    async def get_base_id_in_int(self) -> int:
+        """Gets base id from memory."""
+        mem_line = await self.read_mem_line(1)
+        return int.from_bytes(mem_line[0:4], "big") 
+    
+
+
+
+
+known_objects = [FAM14, FUD14, FUD14_800W, FSB14, FSR14_1x, FSR14_2x, FSR14_4x, F4SR14_LED, F3Z14D, FMZ14, FWG14MS, FSU14, FMSR14, FWZ14_65A, FSG14_1_10V, FGW14_USB, FDG14, FHK14, F4HK14, FAE14SSR, FTD14]
 # sorted so the first match of (discovery name is a prefix, size matches) can be used
 sorted_known_objects = sorted(known_objects, key=lambda o: len(o.discovery_names[0]) + 0.5 * (o.size is not None), reverse=True)